added periodic batching support to the primary sink
- adding support for non blocking batch posts to event grid will ensure the best possible performance
- refactored log event to event grid post code into the new EventGridClient class to better support additional sinks
- bumped minor version

fixes #11
