+++ src/Serilog.Sinks.EventGrid/Serilog.Sinks.EventGrid.csproj
@@ -12,9 +12,9 @@
     <PackageIconUrl>http://serilog.net/images/serilog-sink-nuget.png</PackageIconUrl>
     <RepositoryType>git</RepositoryType>
     <RepositoryUrl>https://github.com/Authenticom/serilog-sinks-eventgrid</RepositoryUrl>
-    <PackageTags>serilog events eventgrid</PackageTags>
+    <PackageTags>serilog sink events eventgrid azure</PackageTags>
     <Copyright>Copyright © Chris Kirby 2017</Copyright>
-    <Version>1.0.3</Version>
+    <Version>1.1.0</Version>
   </PropertyGroup>
   
   <ItemGroup Condition="'$(TargetFramework)' == 'net461'">
@@ -26,6 +26,7 @@
   <ItemGroup>
     <PackageReference Include="newtonsoft.json" Version="10.0.3" />
     <PackageReference Include="serilog" Version="2.5.0" />
+    <PackageReference Include="serilog.sinks.periodicbatching" Version="2.1.1" />
     <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.4.0" />
   </ItemGroup>
